nr-metrics-to-sheets/requirements.txt
requests
python-dotenv
gspread
google-auth
pyyaml